+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 ### Enhancements
 
 - Adds `paywallPreload_start` and `paywallPreload_complete` events.
+- Improves drawer presentation style corner rounding by applying the device radius on bottom corners.
 
 ## 4.11.2